requirements.txt
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyodbc==4.0.39
pytest==7.4.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
Werkzeug==2.3.7
win32-setctime==1.1.0
zipp==3.17.0
